chesstrainer: Create and quiz yourself on chess lines

The purpose of the chesstrainer package is to create sequences of moves (lines), save them, and then quiz yourself on these sequences. This is especially useful for training openings, but can also be used for creating chess puzzles.

Starting the Trainer for the First Time

To start the trainer, run play() from the console (see the play function for some startup arguments that can be adjusted). When starting the trainer for the first time, a default directory is created where sequences will be stored. A few example sequences also come with the package and you will receive a prompt whether to copy these sequences to the sequence directory. To familiarize yourself with the trainer, it would be good to do so.

You will then receive a prompt to enter a player name. Choose a name and hit enter.

The trainer then starts in ‘add’ mode (see ‘Note’ if you receive an error that the graphics device does not support event handling). For now, you should hit <space> (i.e., the spacebar), which will switch to ‘play’ mode. One of the existing sequences will then be chosen at random. The trainer may immediately prompt you to play a certain opening, it may make some moves for you, or it may start at a later position during a game. Either way, you will then have to make the correct move based on the position. You do this by dragging the piece to be played from its current position to the desired target square. The trainer will then make the next move, after which you will then again have to make the correct move. This will continue until the end of the sequence. If you do not know what the correct move is, you can hit the <h> key to get a hint. This will show which piece to move. Hitting <h> for a second time will show the target square.

In play mode, information is shown at the bottom about the current mode, the player name, the name of the current sequence, the current move number and the total number of moves of the sequence, the score (see below), and the number of times that the sequence has been played.

With <F1>, you can receive an overview of all keyboard shortcuts. You can quit the trainer with the <q> key.

The Score

The trainer keeps track of a score for each sequence. For a new sequence, the score starts out at 100. The score is multiplied by 0.80 after completing a sequence (rounded up to the nearest integer). When making the wrong move when playing a sequence, the score is increased by 50 points (up to a maximum of 100). Receiving a hint increases the score by 25 points (50 points when asking for both the starting and target square).

The trainer uses the score to select the next sequence to play. Sequences with higher scores have a higher chance to be presented. The probability of selecting a particular sequence is computed with \(p_i = s_i^e / \sum(s_i^e)\), where \(s_i\) is the score of the \(i\)th sequence and \(e\) is the ‘exponent value’. This is set by default to 2 and enhances the probability that sequences with a large score are more likely to be selected. The exponent value can be adjusted during play with the <^> key. Setting the exponent value to 0 makes all sequences equally likely to be selected. Setting it to a large value (e.g., 100) almost guarantees that the sequence with the highest score is selected next. One can also manually set the score for the current sequence being played with the <o> key. Setting the score to 0 means that the sequence has a zero probability to be selected.

If you make a misclick during a move and want to take the score adjustment back, you can do so with the <t> key.

After completing a sequence, the current score for the sequence is saved (together with the number of times that the sequence has been played) to the sequence file. This way, your progress in learning the sequences is saved permanently across sessions. Note that the scores are player specific.

Managing and Selecting Sequences

Using the <l> key, you can list all sequences, together with their current score, the number of times they have been played, and the current probability of selection. During ‘play’ mode, you can use the <d> key to delete the current sequence. Note that this is permanent and all progress information about the sequence will be lost. You will receive a prompt to confirm the deletion.

Using the </> (or <.>) key, you can select a subset of the sequences. One can select sequences in various ways:

by entering a search term (which is applied as a regular expression to the filenames of the sequences),

by entering a sequence number (e.g., 2 to select the 2nd sequence),

by entering a sequence number range (e.g., 2-5 to select sequences 2 to 5),

one can select sequences with a score above (or below) a certain value (e.g., by entering score >= 80),

one can select sequences that have been played a minimum/maximum number of times (e.g., by entering played <= 1).

Hitting <enter> at the prompt selects all sequences.

Adding New Sequences

With <space>, you can switch into the ‘add’ mode. This is where you can enter new sequences. The first thing to decide is whether the player should play with white or black. With the <f> key, one can flip the board accordingly (the player always plays with the pieces at the bottom). Next, one has to decide whether some moves will automatically be played at the beginning of a sequence. If some moves should be played automatically, then one can simply make these moves on the board. Note that there is no check if the moves are legal and/or are played in the correct sequence, so care must be taken to do so appropriately. Once you arrive at the position where a player should start making moves, hit the <z> key (note that ‘Show’ at the bottom left then turns from ‘Yes’ to ‘No’). Now continue playing the line as it should be played. The sequence should end with a move of the player. If you make a mistake when playing a move, use the <b> key to take the move back. Once the sequence is complete, use the <s> key to save the sequence. You will be prompted for a filename. Once the sequence has been saved, it will automatically become part of the sequences that may be selected in ‘play’ mode.

A sequence may also start immediately in the starting position. Here, you will hit the <z> key right at the beginning. Since it is then essentially impossible for a player to know what they should play (since there are many sensible moves one could make), you will have to instruct the player what they should do. Using the <c> key, you can add a comment to the current move which will be shown to the player. For example, you could tell the player to play the Spanish (Ruy Lopez) opening. You should then make the appropriate moves on the board.

Comments can be added to any move using the <c> key. At times this may be useful if one could venture into multiple sensible lines in a given position. You would then have to instruct the player to make a particular move.

If a sequence should start not at the starting position of a chess game, but later on, one can move the pieces into the desired position and then use the <0> key to make the current position the starting position for a particular sequence. This way, one can create chess puzzles, such as asking the player to find a mate in a certain number of moves. This is currently tedious (since one may have to move many pieces around and removing pieces is only possible by capturing pieces). In the future, the package will include a proper board editor.

When saving a sequence, the color of the pieces to be played is automatically prepended to the filename (i.e., as w_ or b_). This can be useful when selecting sequences (e.g., using the search term ^w_ will select all sequences that start with w_ and are hence sequences where the player plays with white).

Using the <e> key, one can edit a sequence. This is most useful for adding or removing comments in the comment column. Making adjustments to the other columns should only be done if their meaning is clear.

Note that sequences are quite personalized, since they involve going into a particular line that one deems most appropriate. Others may disagree and prefer a different line. Therefore, you should create sequences that you want to train yourself on, so you will be able to play them quickly from memory.

Managing Players

The trainer can be used by multiple players. A player can be selected with the <p> key. One can either enter a player number, the name of an existing player, or enter a new player name. The current player can be deleted with the <r> key. Since this deletes all progress information for this player from the sequence files, you will receive a prompt to confirm this.

With <F2>, a leaderboard (based on the lowest average score) and player statistics are shown (i.e., the average score, the standard deviation of the scores, the minimum and maximum score, and the total number of sequences played). Sequences whose score was manually set to 0 are not included in these statistics.

Sound

The trainer plays various sounds (when playing moves, capturing pieces, when making a wrong move, or when completing a sequence). This may or may not work depending on your operating system and configuration. See the playsound function for further details.

Other Options / Keyboard Shortcuts

When playing a sequence, the graphical elements shown on the playing board are created by repeatedly drawing on top of the board. Therefore, when resizing the board, all elements are redrawn, which takes a few moments. With the <u> key, one can redraw the board in its current state, after which one can resize the board without having to wait for all previous graphical elements to be redrawn.

Clicking the right mouse button on a square adds a circle to the square. Dragging the mouse from one square to another while pressing the right mouse button draws an arrow. Those familiar with the https://lichess.org website will recognize these visualizations. Clicking with the left mouse button on a square, the <u> key, or making a move removes these visualizations.

Various settings can be adjusted while the trainer is running:

By default, there is a 0.5 second sleep time between moves played by the trainer. This can be decreased/increased with the <-> and <+> keys.

By default, the trainer waits after completing a sequence before starting a new sequence. One can toggle this pause on/off with the <w> key.

One can toggle the sound (if it is working) on/off with the <m> key. The sound volume can be decreased/increased with the <[> and <]> keys.

With the <i> key, one can toggle the language of the trainer between English and German (author note: I wrote this trainer in part for my chess playing kids, which is the reason why a German translation is available).

These settings can also be adjusted via the arguments of the play function.

Backing Up Sequences

It is advisable to regularly make a backup of the sequences you have created. The directory where sequences are stored can be found with the command: tools::R_user_dir(package="chesstrainer", which="data").

License Info

The chess pieces used in the package were obtained from https://commons.wikimedia.org/wiki/Category:SVG_chess_pieces where they are available under the Creative Commons Attribution-Share Alike 3.0 Unported License.

The sound files used in the package were obtained from https://github.com/lichess-org/lila/tree/master where they are available under the GNU Affero General Public License.

Note

The trainer makes extensive use of getGraphicsEvent to capture mouse movements and keyboard inputs. Only some graphics devices support this. If you receive the error message ‘The graphics device does not support event handling’, then the default graphics device that was opened does not support event handling. For example, this will be the case for the RStudioGD graphics device that is used by RStudio. You can then try running x11() before starting the trainer with play().

Author

Wolfgang Viechtbauer (wvb@wvbauer.com)