Start the Chess Trainer

Function to start the chess trainer.

play(player="", lang="en", seqdir="", sfpath="", sfgo="depth 20", ...)

Arguments

player

character string to specify a player. If blank, the trainer will prompt for a (new) player name.

lang

character string to specify the language (either "en" for English or "de" for German).

seqdir

character string giving the path to a directory where sequences will be stored. If blank, the default directory is used.

sfpath

character string giving the path (including the executable) to Stockfish. If blank, Stockfish is not used for position evaluations.

sfgo

character string to specify the calculation parameter for Stockfish. Either something like "depth 20" (to fix the search depth in terms of the number of half-moves) or "movetime 1000" (to fix the search time in milliseconds).

...

other arguments

Details

This is the main function to start the chess trainer. See the chesstrainer-package documentation for an introduction to the package.

Author

Wolfgang Viechtbauer (wvb@wvbauer.com).

Examples

# start playing
if (FALSE) { # \dontrun{
play()
} # }